Fix: observe project root not removable directories
It seems that replaced watched directories cannot be observed by
nodemon. The recommended [workaround] is to watch the parent. This patch
changes the files monitored to the package root less ignored files
(.git|node_modules) so that any change has the expected effect.

The output prior to this change was approximately:

  $ DEBUG=nodemon,nodemon:\* npm start
  ...
  nodemon start watch on: dist/client +0ms
  nodemon start watch on: /home/stephen/dev/wmf/marvin/dist/client +0ms
  ...
  clean-webpack-plugin: /home/stephen/dev/wmf/marvin/dist/client has been removed.
  ...

[workaround]: https://github.com/paulmillr/chokidar/issues/555

Change-Id: I5d84a8e65141fe46db9750dcd09ce7c23dfc6600