Chore: convert page template to (P)React
Make the server page template just another component for parity with the
client. `doctype` is specified by the server itself since this is
independent of the component and not practical to include before
rendering. Additionally, fix chunk scripts in the Page component and
update the Page tests. Tests could be made less brittle using Domino and
rendering to and asserting a real DOM.

https://github.com/developit/preact-render-to-string/blob/e1d7f47/README.md#render-jsx--preact--whatever-via-express
https://github.com/facebook/react/issues/1035

Change-Id: Ibe1d18ca2c0e1466146578b50df1afd4fad6c9ab