Chore: Fix npm run build
The build:server step failed because the assets in dist/ were trying to
require a .css file which wasn't present, thus throwing an exception
before the ignore-styles module could do anything.

This patch copies the css files to dist/ as part of the build step so
that require doesn't fail in node, and ignore-styles takes care to make
those requires a NO-OP then.

With this patch applied:

    npm run build && node dist/server/index.js

Should work all fine.

We will consider more sophisticated solutions like using webpack for the
server side building in the future.

Bug: T175255
Change-Id: I52399df4c684dc8261f9bcf74d84a09d5b5c9ba7