New: Add Paper component
As explained in the style guide:
  https://wikimedia.github.io/WikimediaUI-Style-Guide/visual-style.html

Paper is the base abstraction for the visual style of wikimedia. This
patch adds a Paper component and exercises it in the style guide.

Ideally later other page components like (Special)Page or WikiPage will
internally use Paper.

Visit /dev/style-guide to check it out.

Additional changes:
* Added base spacing variable (--space)
* Normalized spacing across App and Paper
* Added a max-width to App to make it a bit more desktop-friendly

Bug: T175804
Change-Id: Ic1841a9558a6df8bd113afe8257d658a4ade1cd2