Chore: Set target value for src/common ts files
Nearest common ancestor is the root tsconfig, so by default we'll set
a modern target for common code, so that we don't get warnings (about
Promise usage for example) when using editors like vscode. We also set
the module to commonjs to avoid dynamic import warnings, and allow
webpack.config.ts to keep working fine.

The entry points, src/{client,server} have their own tsconfig with
different targets, so errors related to the target for their specific
environments will be reported when compiling specifically the client or
server code.

Change-Id: Ib07d51cb6f52bba6e8359ec3c1f74e1477c1a514