Chore: Fix double spacing in Page template
In the Page-content, which has inner bottom spacing due to the spacing
from Paper, the last element's margin could accumulate with the bottom
padding, producing double-space.

As such, make the last element inside the content not have any extrinsic
spacing (margin) that could add to the inner spacing in the Page.

Change-Id: I78635f111caf95ec08e307293269e03b3f630195