Chore: Add Icon component with menu and search icons
Will be used soon in the Header.

This initial version renders icons as inline svgs in order to be able to
style them differently and render them at different screen dpi without
having to generate multiple assets.

In the future we can add parameters to the Icon component, like size,
style, etc to have a flexible icon system.

Bug: T173022
Change-Id: I9b838274e5ab306967062109f5504443e85ae796