Chore: Fix SVG rendering misbehaving on Safari
Both on desktop and iOS, the wordmark SVG specifically would some times
turn the paths into fully filled rectangles.

Apparently, the <base> tag affects the SVG imports with url and others,
breaking them. Sources:

* https://stackoverflow.com/a/18265336/1400662
* https://github.com/bodymovin/bodymovin/issues/360
* https://www.w3.org/2015/08/25-svg-minutes.html#item08
* https://github.com/angular/angular.js/issues/8934

This change removes the <base> tag, and fixes the production
configuration for webpack so that the assets build and are loaded fine
on nested paths.

The issue was that absolute paths were used implicitly on manifest.ts
by adding a "/" to the assets-manifest.json entries, without webpack
knowing about it. As such, when trying to load assets dynamically
(dynamic imports) webpack wouldn't know about the "/".

That root marker has been moved to webpack's configuration.publicPath
when compiling for production, so that the assets manifest includes the
proper root "/" and the dynamic import loading mechanism also knows
about it.

Tested in development and production mode, by loading the style guide
for example both from the server, and from a different page navigating
to it on the client.

Bug: T173022
Change-Id: I432e549e24659f38171fc45c8b2d29cdef88b638