Bug: Fix Link not capturing certain clicks
When clicking on DOM elements inside the anchor element, the Link
component wouldn't capture the click and direct it to client side
navigation.

This is because it was using event.target internally, when that is the
originator of the event. Instead, we want to check the anchor, which is
the currentTarget instead, the element on the current phase of bubbling.

This patch fixes the problem and adds a comment.

The bug can be experienced by clicking on the Wordmark in the header,
where the link contains an SVG element and the event.target would end up
being a path element or something else.

Change-Id: I156d9c293ecae85c05af5d6c2112e1d6236c76c9