Chore: Automatically generate the vendor bundle
Instead of having to manually specify the vendor bundle which is prone
to mistakes and to get outdated, automatically generate the vendor
bundle with the files included from node_modules/ without including the
styles for now.

It may be interesting to add support for a vendor.css at some point if
we use any big library.

Other changes include moving the use of 'preact/debug' to the code
itself, and relying on dead code elimination to be removed in the
production version.

Also updated some of the comments regarding the CommonsChunkPlugin
usages.

Change-Id: I0d66d69efdd60fd911cf1ed4bdf35d30019cd064