Perf: Avoid style-loader in dist/public/index.*.js
Webpack was including style-loader in the bundle so that it was able to
load style modules on lazy loaded code (as the fallback in
ExtractTextPlugin in the loaders), which it needs when allChunks is
false.

We are including all CSS in the same base file, so it isn't needed, so
this patch sets allChunks to true.

See docs at https://github.com/webpack-contrib/extract-text-webpack-plugin

   options.fallback	{String}/{Array}/{Object}	loader(e.g 'style-loader') that should be used when the CSS is not extracted (i.e. in an additional chunk when allChunks: false)

This saves 5kB from the bundle.

Change-Id: I6455715c9cddb4ff3ae221aba209588bbc57d151