Update: add initial baseline content styles
This patch adds initial content styling based on previous work in Loot
UI and Weekipedia. Numerous tweaks were made based on testing and
unusually nuanced specifications were replaced with approximations from
the root which are probably also incorrect but at least do not
needlessly incur tech debt. In general, all rules that could not be
found to be effective were removed.

Where possible, examples of rule application are included with page
revision so they can hopefully be maintained. All URLs are Marvin
development-specific since comments are generally for developers
actively developing.

Lots more work will be needed but not as part of this task. Extension
specific styles are not included in this patch but rules for the
Plainlist template were copied verbatim.

https://github.com/joakin/loot-ui/blob/adc6a04/lib/client/components/section/section.less
https://github.com/jdlrobson/weekipedia/blob/9228a2f/libs/client/components/SectionContent/styles.less

Bug: T173330
Change-Id: I96f18897dabebd9fc6b1bb1bd2935d5098b37d39