Update: add popular external styles
Add core gallery, Parsoid, Cite, Math, and Timeline extension styles.
These styles are of external origin and were copied from their sources
(noted in the header of each file). These specific styles were chosen
for an initial offering because they are bundled with the Android app
but do have tradeoffs of complexity, size, and maintainability for
inclusion or exclusion.

https://phabricator.wikimedia.org/diffusion/APAW/browse/master/scripts/make-css-assets.bash;3718d876dc0eee7711680cdf6924dbdd90e5c9cb$9

Bug: T173330
Change-Id: Ib2833682daeafbc55c2144265145e3d64fe3f53d