Chore: fix ESLint max length ignore directives
`// eslint-disable-line max-len` disables the max-len rule for the given
line. However, this rule also included an `ignorePattern` for any line
with an ESLint directive. This caused the rule to be disabled AND the
line to be ignored which triggered --report-unused-disable-directives.
Remove the ignore pattern for max-len specifically so only the rule will
be disabled and the line can be checked accurately for disable directive
usage.

Change-Id: I154335c7504dc0e73339acb285e9842291a858d5