Update: report server URL redirect statuses
When a redirect occurs on a PageModule.getInitialProps() request issued
server-side, such as
https://en.wikipedia.org/api/rest_v1/page/summary/Albert%20Einstein,
report the 3xx HTTP status instead of following the URL. This gives
browsers the expected behavior when visiting a non-canonical URL. No
changes were intended to client-side behavior yet.

PageModule.getInitialProps() requests work the same as before when no
redirect occurs. However, when a redirect does occur on the server, this
is considered exceptional and an error with the status code and resolved
URL is thrown. For example, a getInitialProps() request for
https://en.wikipedia.org/api/rest_v1/page/summary/Albert%20Einstein
would throw an error with an HTTP status of 301 and a URL of
https://en.wikipedia.org/api/rest_v1/page/summary/Albert_Einstein. The
PageModule itself catches the error and constructs the correct
destination URL for that page, since only it knows how to get the route
parameters needed to pass to Route.toPath(), for example,
http://localhost:3000/page/summary/Albert_Einstein, then finally
rethrowing the error with the new URL for the server router to handle.

An alternative implementation for redirected responses was considered:
instead of throwing an error, return a response with an empty body. This
might work well later when TypeScript discriminated union types become
more fully featured. However, today this would percolate response body
existence checking at every layer. The unmarshaller has nothing to do
with a redirected response so this seems unwanted. Additionally,
throwing an error is not entirely beyond compare in that the Fetch API
itself supports a similar redirect behavior when RequestInit.redirect is
set to "error".

It would be desirable for the server to handle all redirect logic in a
generic way for all requests but it's not clear how to do that. The
getInitialProps() RESTBase URL needs to be destructured and a new Marvin
URL must be created from the result.

Bug: T177681
Change-Id: I89e03f7a9e45f318c01e26379114a195ebbff2cb