Update: use CSS border-box box model
We don't yet know what all the implications of this change will be
across wikis but we want to start seeing any issues that exist sooner
rather than later. This patch adds the border-box box model for all
elements but may later be reverted or selectively disabled for wiki
content if found to be problematic.

Bug: T176999
Change-Id: I443658458da2eaa3e945de886becf6bbd71d3e03