Chore: replace clean script with Webpack plugin
Remove the clean package script and add the clean plugin to the Webpack
config. The build process will be entirely owned by Webpack for server
and client builds so cleaning can now be part of the config.*  The
tradeoff is that Webpack will be able to more fully own the build
process and run as a single command instead of clean then build but at
the expense of an added NPM dependency.

*Note: this patch will temporarily allow a dirty server build directory
but it will be fixed when the server uses Webpack for building.

Bug: T177235
Change-Id: I4b91c492e52bac35834a8160791bdfc0ac2f1822