Chore: don't restart client on server changes
It's true that the client depends on the server. However, the dependency
is limited in scope to an initial server-side rendered HTML page and any
preloads including the client itself. In general, it does not make sense
to restart the client for a server change since an SSR is really what
needs to be triggered. If this functionality can be reliably provided
using a unified Webpack config, we may restore it later.

Bug: T177235
Change-Id: I294fa894e1dcf294c04150b9e876a6858095ec9e