Hygiene: Use AnyComponent from Preact always
With the previous patch to update to typescript 2.6 with strict function
types, we had to reintroduce our own AnyComponent definition because our
FunctionalComponents were causing troubles with the preact type
definitions on the PageComponent definition.

Instead of having around two definitions of AnyComponent floating
around, this patch uses the preact ones everywhere, but specifies on the
PageComponent type that the Component can be partial, so that our
FunctionalComponents fit in.

Change-Id: I4b51047078c9abe7bebe2286e606c870954bcd31