Fix: don't reference the manifest in the client
The client cannot access the asset manifest because its generated from
the client itself. Move manifest read logic back to the server but keep
the asset path generation logic in common. This patch fixes a broken
manifest require reference in the client and unintentional modification
of assets#scripts in the server.

Bug: T180893
Change-Id: Iebb827f597575932e87dd1f0e23673a92eae7f7b