Fix: image links don't load destination in place
Check the parent of clicked images to see if they're an anchor. If so,
render on the client. This change fixes the lead images but may be found
to fail for other scenarios. However, it's consistent with the
preexisting implementation of checking event.target. If this change
proves inadequate, event.target should be ascended until an anchor is
found.

Bug: T181901
Change-Id: Ia2a50555675c93cb3784d271f8f9bcd7ccff4c80