Update: change revision path param to oldid query param
Change the URL path parameter for wiki page revisions to a query
parameter. e.g.:

  http://localhost:3000/wiki/Film/803321621#Notes

  ↓

  http://localhost:3000/wiki/Film?oldid=803321621#Notes

This change means that pages with previously ambiguous titles are now
well-defined. e.g., 0/0 is known to be about indeterminate form and not
the number zero at revision zero. Additionally, all slashes, even
trailing, are now considered part of the title.

This change was deemed necessary for API compatibility with existing
links. Outdated references in the code were updated with:

  find src -type f|
  xargs -rd\\n sed -ri 's%(localhost.*)/([0-9]+)%\1?oldid=\2%g'

Bug: T178616
Change-Id: If9220083f64c35afba105a01858167643076014d